requirements.txt
fastapi
pydantic
uvicorn
mysql-connector-python
pandas
python-dotenv
python-multipart